Adds a "dynamic" module to dynamically load services from smithy modules
This relies on mutability + suspension to build a map of schemas that
work against dynamic data, backing product types with arrays.

The goal is notably to get to a state where services could be proxied
dynamically, whether for actual proxyfication purposes, or to create
CLIs that do not need recompilation, or create UIs/playgrounds
dynamically.