Change DynamicModelCompiler to rely on `cats.Eval` instead of schematic suspension
* Add cats as a dependency to the dynamic module
* Change DynamicModelCompiler to avoid relying on schematic suspension,
which creates a burden on the Schematic implementation for awareness
of the ad-hoc usecase that is dynamic schema. Instead, we rely on
`cats.Eval` to defer the construction of schemas to *after* the map
of shapes has been traversed.

This is a trade-off : the `Eval` get unwrapped at the end of the
compilation, forcing their evaluation, which might have some impact
on the dynamic CLI's snappiness.