Best effort error parsing
Best effort parsing happens if:

* no errorTypeHeader is found to discriminate
* status code matches multiple error type or no error type

In this case, the following logic happens:

* we try to decode for every union alternative
* first success wins
* order is "smart" (see below)

Ordering of the alternatives depends on the status code. If status
is in the 4xx, then generic client error are placed first, otherwise
it's server ++ client ++ the rest.

See https://awslabs.github.io/smithy/1.0/spec/core/http-traits.html#httperror-trait

Note: if one of the alternative has only optional event, chances
are that parsing will be successful