Schema overhaul (#140)
This is an overhaul of the "schema/schematic" abstraction to make use of a GADT. We're trading off extensibility for ease of implementation for interpreters. In particular, having a GADT at hand allows us to peek into the structure of a schema (and the hints carried at each layer), several layers down, without a need for ad-hoc constructs such as `Hinted` and such. 

It also allows interpreters to delegate to other interpreters when convenient. 

It also adds the notion of `surjection` in the schema, which allow to capture transformations that work only on a subset of the type of the schema. The usefulness of it has to do with making it easier for downstream implementors to incorporate validation concerns within the decoders. 

Additionally, traits that imply validation of the incoming data (range/length/pattern) are now receiving specific treatment during codegen in order for the schemas containing them to leverage the surjection. 

Co-authored-by: Jakub Kozłowski <kubukoz@gmail.com>
Co-authored-by: David Francoeur <dfrancoeur04@gmail.com>