Add options for rendering case class ADT members directly within union (#211)
Adds a way for users to signal that a structure should be considered as an ADT member of a union. This allows to eliminate one layer of wrapping that often feel superfluous. 

This also prevents a structure from being referenced anywhere else but the union it is a member of. This is to not create a situation where several case classes would map to the same smithy shape, which would be counter intuitive and break least surprise. 

This PR includes:

- The smithy definition for the adtMember trait
- A validator to make sure all usages of the adtMember trait are valid
- Updates the renderer to pay attention to and respect the adtMember trait