Couple ShapeTag with Schema
Adds a constraint that enforces that all shape tags have an associated
schema. This allows to completely remove the need for `SchemaIndex`, and
greatly simplifies the dynamic compilation of models.

The simplification is allowed by a change to the Hints implementation
which can now store two kinds of bindings : static ones and dynamic ones.
Dynamic bindings are essentially a pair of ShapeId and Document.
Because ShapeTags now have an associated schema, this schema can be used
to deserialise the Document on the fly, which means that the Dynamic
compiler does not need to receive schemas in advance anymore.

The cost is that some manually defined shapes need to receive a Schema
now.