Rework of the concept of collection in the Schema abstraction (#290)
This replaces list/set in Schema by a generic collection construct, that takes a `CollectionTag` as a parameter. The `CollectionTag` carries methods to (de)construct the collections, and can be used to generically implement SchemaVisitors without worrying about what collection we're dealing with. In addition, the `CollectionTag` is sealed, and can be inspected to provide specialised version of the encoding/decoding logic when relevant. 

Co-authored-by: Olivier Mélois <olivier.melois@disneystreaming.com>