Scala native support (#289)
Adds Scala Native support (for Scala 3 only, to avoid the matrix increasing in size too much for something that is quite bleeding edge). 

The supported modules are : 

* smithy4s-core
* smithy4s-json 
* smithy4s-dynamic 

This PR also amends the dynamic module slightly to make it depend on the smithy-model library (on JVM only), facilitating the loading of dynamic services/schemas directly from a smithy Model. This makes tests a little bit easier to write and reduces the need for sharing some test-utilities across modules, which was leading to a cyclic module dependency that I was too lazy to solve via yet another indirection. 

Additionally, because smithy4s-dynamic is built on native now, I had to remove the dependency to CE/weaver in its test. I wanted to keep the diff smallish, so I ended up making a dummy IO construct (totally not pure) to reduce the noise a little. When CE builds on native, we'll switch back.