Revises the "union" method in SchemaVisitor ... (#302)
... to make it easier to write memoised dispatching logic

The `union` method in SchemaVisitor is an obvious footgun, because it makes it a hard to memoise the compilation of schemas for each alternative. 

This PR revises the `union` method to expose a new construct `Alt.Dispatcher`, which does the heavylifting of memoizing the result of compiling schemas, and dispatching the union instance to the correct alternative. 

* Creates a `Alt.Dispatcher` construct that aims at streamlining the
implementation of the `union` methods in SchemaVisitors, which is the
hardest due to the fact that the implementor has to think about memoization.
* Creates an `EncoderK` typeclass (used by Alt.Dispatcher) to generalise
over what is an encoder, and streamline further the implementation of
union encoders.
* Examplifies the use of `Alt.Dispatcher` in DocumentEncoderSchemaVisitor

This should remove the need for https://github.com/disneystreaming/smithy4s/pull/260