Protect against models colliding with user-facing types
Adds logic to protect against rendering types using (non-fqn) names that
Scala devs are bound to use when using smithy4s. (stdlib, smithy.Timestamp, etc).

For instance, when encountering `foo#List`, then a `_List` scala type gets created
instead.

If `List` is the name of an operation, then the method created is `list`, in an
effort to minimise the impact on user experience.