Facilitate sharing specs across multi-module builds.
This adds necessary the necessary mechanisms and logic to facilitate
sharing smithy specs in multi-module settings.

The first important part is the addition of a resource generation
mechanism that adds the local smithy files to the output resources,
and synthetise smithy files containing only metadata, keeping track
of the namespaces that have been generated by smithy4s.

Because the metadata in question is an array node, it is monoidal in
nature and the values are aggregated when loading models. These values
are used to skip code generation in downstream Smithy4s calls, as the
downstream modules can trust that the upstream one will have generated
the relevant package.

Then, the SBT plugin changes to load internal dependency outputs as
local jars. This has the effect of pulling the models from upstream
modules in the current model assembler, allowing users to depend on
the shapes it contains.