Amends cache key in Cached schema visitor
In order to reduce risk of memory leaks when people do not have stable
references for the routers (ie `def` instead of `val`), this PR changes
the cache key for the Cached schema visitor to avoid it being impacted
by bits that are guaranteed to be different with respect to
hashing/equality, such as lambdas (accessors, thunks passed to Lazy,
etc).

This isn't a panacea, Schemas cannot be accurately compared for equality,
for the sheer fact that two schemas might be semantically the same
without being equal, due to the language limitation of not being able
to compare lambdas.

Therefore, we mitigate the problem by providing a custom hashing
function for schemas, and use that (conjointly with the ShapeId) as a
key during memoization of schema compilation.

Additionally, this PR provides an implementation for the equals/hashcode
methods of the Hints map.