Amends the caching mechanism to avoid relying on global variables
This is changing the SchemaVisitor caching mechanism to allow to inject
a mutable cache rather than relying on global ones, which can lead to
memory when users apply unpredicted usage patterns.

Injecting a mutable cache means we can scope the caching to the
lifecycle of interpreters, which is a lot safer.

The problem really stems from the fact that Schemas are not stable (with
regards to hashCode/equality) methods when re-instantiated or when
undergoing hint transformations, therefore cache miss are likely to
occur even when semantically a cache hit should happen.

Additionally, this changes the Dynamic module to try and yield
as-stable-as-possible schemas to further reduce the risk of memory
leaks whilst benefiting from the caching mechanism.