Client httpReponseTest for output
I'm uncertain about the approach I've used here. Implementing test
response case for client without the server interpretor (and vice-
versa for the client) is tricky.

For example in this PR, where only client tests are available. I have
to grab the codecAPI to get the mediatype and set it on the response
(if not provided in the test case).

On the server side it's worse because I have to craft a request out of
thin air. Previsouly I had managed to do it because to have compliance
tests we required both the client and server side of the protocol
to be implemented. But because, we split them, it's much harder.

I still have trouble to put in writing how I struggle with this feature lol.

Anyway I'll drop what I wrote in markdown in a comment on this PR, let me
know if it makes sense