[series/0.17] fix CI trigger (and build) (#573)
* Fix glob pattern in CI trigger

* Disabling jar cache in tests

* Streamline fork setting

* Avoid assuming dependencies during codegen (except for alloy)

* Only dependency assumed is alloy-core
* Assumption happens via build-tools default settings being set by
plugins/modules
* If users want to generate code from AWS, they need to manually
ascribe the dependencies containing AWS traits (updated the documentation in consequence)

Alloy can evolve independently from Smithy4s, and we cannot assume the
files it contains will remain at stable locations across versions, so
hardcoding a reference to the smithy files contained by alloy artifacts
would be ill advised. Likewise, we cannot force the use "discoverModels"
because it can backfire easily. The only solution is to have the plugins
force the fetching of alloy (by default) at the build-tool level.

* Update .scalafmt.conf

* Regenerated core

* Correct reference in documentaiton

* Amend model loading logic and fix plugins

* Amends the model loading logic to avoid using discoverModels when
the user doesn't specify it explicitly. It turns out classloaders are
not a great way of achieving loading, and that using the ModelDiscovery
construct manually is better.
* Fixes the `smithy4sTransitiveIvyDeps` call so that downstream modules
do not have to re-iterate the dependency declaration
* Fixes the SBT plugin so that library dependencies with `% Smithy4s` are
transitively acquired
* Adds some tests

* Regenerate Scala 3 boilerplate

* cleanup

* Headers and formatting

* Fix docs

* Revert param name