Change the build plugins so that libraryDependencies are taken into account for codegen
Now, by default, any library declared as a "normal" dependency, by means
of `ivyDeps` in mill or `libraryDependencies` in sbt are used during the
code-generation process.

This makes Smithy4s more consistent with itself, as the same behaviour
is applied to internal dependencies.

It is also argueably more intuitive to do it this way.