[SBT] : unifies resolution before codegen
Unifies resolution (via SBT's resolver) to ensure that there are no
conflict resolved jars that are passed to the code-generator. This is
done by gather all ModuleIDs before re-running the resolver, in order
for it to perform the necessary eviction.

Thus, we only concatenate the jars that come from dependency resolution
with the internal jars (which should hardly ever conflict).