Changed generated implementation of some Service methods
In situations where services has a LOT of operations (typical for AWS
services), the Scala compiler can actually crash with a "method to long"
error, as we were generated pat-mats.

This new encoding avoids reliance on pat-mats. It may actually have a
nice side effect of reducing compilation times on the generated code