Fixes Scala code being copied in resources folder
Under the right condition, before this fix, generated scala code would
be copied to the resources folder, which would lead to some compilation
errors due to duplication, because the compiler picks up resources as
sources for some reason.

The bug was introduced in https://github.com/disneystreaming/smithy4s/pulls?page=2&q=is%3Apr+is%3Aclosed,
which generate some smithy files under the `sourcesManaged / "smithy"` directory,
which is then used as an input for the codegen. The codegen has a resource
generator to it, which copies local files under input directories to the
META-INF/smithy resource folder.

However, when generating a namespace that starts with "smithy", the
Scala code gets generated in the same sourcesManaged / "smithy" directory.
Arguably, the problem is that it should be generated in sourcesManaged / "scala" / "smithy",
but at this point I'm unsure what the implications for doing that are. So instead,
I'm ensuring that whatever gets copied in META-INF/smithy are actual smithy files.