Http interpreter reworks
* Creates a `Metadata.Decoder` that will entirely replace the partial
decoders in the future. They co-exist in this PR to reduce the size
of the diff. The reason `Metadata.Partial` is being removed is because
the concern of data-partitioning is moving away from decoders, and
towards the schema.

* Creates a `http4s-kernel` module containing granular abstractions
for capturing the Serde concerns that the http interpreter has to work
with. The `EntityDecoder` and `EntityEncoder` abstractions provided
by Http4s are not suited for our usecase because they are specialised
for http-bodies.

* Reworks error selection logic to make it more granular, which will
allow for usage in the AWS context in a later PR.

* Reworks the http4s module to make use of the new abstractions. This
significantly reduces the coupling between the ServerEndpoint/ClientEndpoint
implementations and the intricacies of reconciliating bits of data that
come from different place. That particular logic is now owned by the
`http4s-kernel` module.