Account for TCP routing when testing how routers decode requests.
Avoid solely relying on intercepting IntendedShortCircuit exceptions
when testing routers, as this method doesn't account for routers that
exercise the TCP layer.

Instead, we still try to intercept that exception (which will work for
routers built on top of Http4s, or routers that do an in-memory conversion
to the Http4s models), but we now also try to intercept 5xx errors before
waiting on a promise. A Router exercises the TCP layer will be likely
to turn uncaught exceptions (like the one we raise) into a 500 http
response.