Initial commit for restJson feature branch.
* Fixes expectations in compliance tests to avoid extrapolating. Headers,
which are particularly tricky considering how AWS runtime libraries work,
are now asserted by collapsing them into a single `Map[String, String]`.
This implies that AWS-specific interpreters HAVE to collapse the headers
themselves in order to abide by AWS's weird encoding of http-date timestamps.
* Removes PUnit from the list of primitives, in favour of a struct-based
representation of the `Unit` type, which is more faithful to Smithy.
* Fixes the SimpleRestJson to pass the RestJson1 tests that have to do
with servers being expected to send empty jsons even when all the output
fields are bound to http metadata.

The initial commit of this branch does not attempt to fix any behaviour
related to AWS interpreters currently not passing tests. It just acts
as a baseline.

Co-authored-by: Yisrael Union <ysrlunion@gmail.com>