Fix content-encoding test on aws
There's a test [that set a custom content-encoding](https://github.com/awslabs/smithy/blob/780b4d9077456b5985db7537fcaebe19b67283f8/smithy-aws-protocol-tests/model/restJson1/requestCompression.smithy#L292).

It ensures that if a user sets a custom content-encoding via
the httpHeader trait, and has requestCompression on the operation
then the final header is akin to "custom value, gzip". The original
implementation did not do that.

I've fixed it here, let me know what you think of that fix. If it's
appropriate, I'll bring it over to the other branch in #1019