Start overhauling of Field.
Turn the Field GADT into a case-class. The concept of optionality
is now handled by `Schema.NullableSchema` (which may be renamed to
`OptionSchema`).

The requirement for adding `NullableSchema` came from having to handle
cases like `sparse` collections, where the nullability is not
contextualised to a field. Unfortunately, this addition means that there
are two ways of defining optionality : one directly on the schema, one
on the field.

Fortunately, it appears we can eliminate this duplication by making the
`NullableSchema` the only way of modelling optionality. This seems to have
the very nice quality of simplifying a lot of the code, at the cost of
changing the `Field`'s interface in order to make the concept of `default`
easy to access. A `NullableSchema` simply yields default values equal to
`None` when the underlying schema has no default value.

**Additionally** : this recovers the distinction between field-level
hints and schema-hints. So far, the lack of distinction had not been a
problem, but the lack of it is detrimental to the implementation of valid
Xml codecs.