SimpleRestJson empty optional payloads == null
Changes the logic so that the default PayloadCodec compilers do not
take the responsibility of turning `null` to `{}` as this is a protocol
specific concern.

Moreover, in SimpleRestJson (which literally has all payloads be json),
it makes sense that an optional payload field set to empty should be
null. This behaviour differs from the case of literal empty
inputs/outputs, where the body should be eluded.