Flatten AWS newtypes
AWS in their smithy specs have a tendency to create newtypes for
the smithy standard shapes. As part of this change, introduce a
projection transformer that will run as a preprocessor before code
generation and will replace the newtypes (if they match some criteria)
with references to the standard smithy shapes.

For a newtype to be considered as being "replaceable" it has to match
the following criteria:
- doesn't have any traits (other than @default)
- is a simple shape
- the name matches the smithy shape (one exception is Date, as it
will be replaced with Timestamp even though it doesn't actually match).
- comes from an AWS namespace (i.e. starts with com.amazonaws)