Optics Rendering (#1103)
Adds the ability to render optics (Prisms and Lenses) in the generated code. The idea is that there will be very basic optic capabilities built into the smithy4s.core and then if users want something more advanced they can use implicit conversions to use something like monocle (see docs in the PR for more info).

Currently Lens instances are generated for structures and Prisms are generated for Unions and Enums.

For newtypes, I added a .value function on prisms/lenses that uses the bijection to automatically create a new lens/prism that targets the inner type of the newtype. I opted for this because then users can choose to create lenses/prisms that target the newtype or the inner type depending on their preferences. Lmk if you think there is a cleaner approach though.