Remove TODO re. PayloadCodec
PayloadCodec is used for Jsoniter based reading/writing because Jsonitor
has a single trait for the read and write behaviour, and that's why
PayloadCodec is used in those code paths. But for cases like this where
reading and writing code paths are relatively separate, there's no
benefit to combining them in a PayloadCodec.