Changes CachedSchemaCompiler to cache the outer schema.
The CachedSchemaCompiler.Impl class suffered from a pretty dire problem
of only caching what is passed to the schema visitor. However, schemas
are often pre-processed before reaching the schema visitors, which,
combined with a lack of guarantee on Schema equality (in particular for
Enumeration Schemas in 0.18), and global caches posed a risk of OOMs in
some cases.

This change makes it so that the initial schema passed to the compiler
acts as a cache key to a cache that's separate to the one that is
typically passed to schema visitors.