Add nullable support to JSON parsing
The JSON codec will now check for a Nullable hint on the schema it is
parsing and not immediately parse fields as None if present. Note that
getting this to work with a new Nullable type does require adjusting
the code checking whether a Schema is for an Option a little so that it
can also include bijections from an option type.