Adjust codegen code to handle Nullable
This involves totally reworking the Field.Modifier logic, as the
addition of Nullable makes the number of cases too complex to keep track
of in an enum alone: the modifier logic now depends on
* whether there is a @required annotation present
* whether there is a @nullable annotation present
* whether there is a default value present, and (for nullable) whether
this value is null or a concrete value

making for a total of ten theoretical cases resulting in four possible
types in the Scala code. We model this in a case class and attempt to
shift some of the logic, such as ordering or working out the wrapped
type, to the class itself in order to avoid other parts of the code
needing to work with its internals.