Add nullable sample spec and smoke tests
There is a mild tension here: on the one hand, we really want to test
all the edge cases of different field modifier combinations to make sure
the generation code works as wel expect. On the other, these combinations
generally have only an extremely limited use or are just overly convoluted
ways to specify the default behaviour, and so they shouldn't be in the
sample specs people use to check usage. We handle this by providing two
test cases: Patchable, which handles the merge patch behaviour this whole
change is meant to support, and PatchableEdgeCases which is explicitly
described as only there for testing purposes.