Add support for @nullable trait in order to support MergePatch usecases (#1408)
## Add nullable support to JSON parsing

The JSON codec will now check for a Nullable hint on the schema it is
parsing and not immediately parse fields as None if present. Note that
getting this to work with a new Nullable type does require adjusting
the code checking whether a Schema is for an Option a little so that it
can also include bijections from an option type.

## Adjust codegen code to handle Nullable

This involves totally reworking the Field.Modifier logic, as the
addition of Nullable makes the number of cases too complex to keep track
of in an enum alone: the modifier logic now depends on
* whether there is a @required annotation present
* whether there is a @nullable annotation present
* whether there is a default value present, and (for nullable) whether
this value is null or a concrete value

making for a total of ten theoretical cases resulting in four possible
types in the Scala code. We model this in a case class and attempt to
shift some of the logic, such as ordering or working out the wrapped
type, to the class itself in order to avoid other parts of the code
needing to work with its internals.

## Add nullable sample spec and smoke tests

There is a mild tension here: on the one hand, we really want to test
all the edge cases of different field modifier combinations to make sure
the generation code works as wel expect. On the other, these combinations
generally have only an extremely limited use or are just overly convoluted
ways to specify the default behaviour, and so they shouldn't be in the
sample specs people use to check usage. We handle this by providing two
test cases: Patchable, which handles the merge patch behaviour this whole
change is meant to support, and PatchableEdgeCases which is explicitly
described as only there for testing purposes.

## Update dynamic module

It's acceptable for fields with a default value to be wrapped in Option
in the dynamically generated code, and trying to align it with the
static codegen behaviour is causing problems in the compliance tests.
We remove that change and any tests that relied on it.

## Add docs for @nullable