Switch out uPickle for Circe in codegen
This is a preliminary change for being able to support smithy-build.json as
input configuration objects; parsing these in a future-proof way is relatively
complicated and difficult to support in uPickle without hand-writing the entirety
of the parsing logic. Since Circe is already used elsewhere in the project, we
replace the uPickle with Circe, including the implementation of creating the
smithy-build.json file from settings.

Note that the special requirement of merging two JSON files while concatenating and
deduplicating their arrays is not supported out of the box by Circe's deepMerge
and still has to be handled manually (cf https://github.com/circe/circe/issues/271)