Use set for repos, imports and dependencies
The reason we saw duplicates is because we have multiple
modules and they all inherit from the project plugin that
is enabled. Because we use `Seq`, we just keep on adding new
elements. Using a `Set` ensure we only have uniques in the end.

When running the command twice, the duplicates are cleared. This
happens because the deep merge algorithm calls `distinct` when it
merge arrays.